[FEATURE] Send HTTP header 'Authorization'
When a server requires an 'Authorization' HTTP header,
the request to fetch the 404/403 page must contain
the same header too to be succesful.
In case of using HTTP digest authentication, the credentials
must be configured in extension manager or sys_template
record. Additionally cURL must be available and enabled.

Change-Id: Iea4d5723b3fbac1186a5453f07cbcd0f77608799
Resolves: #72868
Releases: master
Reviewed-on: https://review.typo3.org/46135
Reviewed-by: Christian Futterlieb <christian@futterlieb.ch>
Tested-by: Christian Futterlieb <christian@futterlieb.ch>